+++ HPXMLtoOpenStudio/resources/constructions.rb
@@ -874,19 +874,19 @@ class Constructions
     apply_window_skylight(runner, model, 'Skylight', subsurface, constr_name, ufactor, shgc)
   end
 
-  def self.apply_partition_walls(runner, model, constr_name, mat_int_finish, frac_of_ffa,
+  def self.apply_partition_walls(runner, model, constr_name, mat_int_finish, partition_wall_area,
                                  basement_frac_of_cfa, cond_base_surfaces, living_space)
 
     imdefs = []
 
     # Determine additional partition wall mass required
-    addtl_surface_area_base = frac_of_ffa * living_space.floorArea * basement_frac_of_cfa
-    addtl_surface_area_lv = frac_of_ffa * living_space.floorArea * (1.0 - basement_frac_of_cfa)
+    addtl_surface_area_base = partition_wall_area * basement_frac_of_cfa
+    addtl_surface_area_lv = partition_wall_area * (1.0 - basement_frac_of_cfa)
 
     if addtl_surface_area_lv > 0
       # Add remaining partition walls within spaces (those without geometric representation)
       # as internal mass object.
-      obj_name = "#{living_space.name} Living Partition"
+      obj_name = 'partition wall mass above grade'
       imdef = create_os_int_mass_and_def(model, obj_name, living_space, addtl_surface_area_lv)
       imdefs << imdef
     end
@@ -894,7 +894,7 @@ class Constructions
     if addtl_surface_area_base > 0
       # Add remaining partition walls within spaces (those without geometric representation)
       # as internal mass object.
-      obj_name = "#{living_space.name} Basement Partition"
+      obj_name = 'partition wall mass below grade'
       imdef = create_os_int_mass_and_def(model, obj_name, living_space, addtl_surface_area_base)
       cond_base_surfaces << imdef
       imdefs << imdef
@@ -907,8 +907,8 @@ class Constructions
                          Material.AirFilmVertical)
   end
 
-  def self.apply_furniture(runner, model, mass_lb_per_sqft, density_lb_per_cuft,
-                           mat, basement_frac_of_cfa, cond_base_surfaces, living_space)
+  def self.apply_furniture(runner, model, mass_lb_per_sqft, density_lb_per_cuft, mat, cfa,
+                           basement_frac_of_cfa, cond_base_surfaces, living_space)
 
     # Add user-specified furniture mass
     model.getSpaces.each do |space|
@@ -948,17 +948,17 @@ class Constructions
       imdefs = []
       if space == living_space
         # if living space, judge if includes conditioned basement, create furniture independently
-        living_surface_area = furnAreaFraction * space.floorArea * (1 - basement_frac_of_cfa)
-        base_surface_area = furnAreaFraction * space.floorArea * basement_frac_of_cfa
+        living_surface_area = furnAreaFraction * cfa * (1 - basement_frac_of_cfa)
+        base_surface_area = furnAreaFraction * cfa * basement_frac_of_cfa
         # living furniture mass
         if living_surface_area > 0
-          living_obj_name = mass_obj_name_space + ' living'
+          living_obj_name = mass_obj_name_space + ' above grade'
           imdef = create_os_int_mass_and_def(model, living_obj_name, space, living_surface_area)
           imdefs << imdef
         end
         # basement furniture mass
         if base_surface_area > 0
-          base_obj_name = mass_obj_name_space + ' basement'
+          base_obj_name = mass_obj_name_space + ' below grade'
           imdef = create_os_int_mass_and_def(model, base_obj_name, space, base_surface_area)
           cond_base_surfaces << imdef
           imdefs << imdef
@@ -978,7 +978,7 @@ class Constructions
     # then the user should input twice the area when defining the Internal Mass object.
     imdef = OpenStudio::Model::InternalMassDefinition.new(model)
     imdef.setName(object_name)
-    imdef.setSurfaceArea(area)
+    imdef.setSurfaceArea(UnitConversions.convert(area, 'ft^2', 'm^2'))
 
     im = OpenStudio::Model::InternalMass.new(imdef)
     im.setName(object_name)